+++ src/components/forums/topic/PostContent.tsx
@@ -83,13 +83,9 @@ export function PostContent(props: PostContentProps) {
   const onReplyToPost = async () => {
     setSendingReply(true);
     try {
-      const tx = await forum.replyToForumPost(
-        post,
-        forumData.info.collectionId,
-        {
-          body: reply,
-        }
-      );
+      const tx = await forum.replyToForumPost(post, forumData.collectionId, {
+        body: reply,
+      });
       setSendingReply(false);
       setShowReplyBox(false);
       setReply("");
@@ -125,7 +121,7 @@ export function PostContent(props: PostContentProps) {
     try {
       const tx = await forum.deleteForumPost(
         postToDelete,
-        forumData.info.collectionId,
+        forumData.collectionId,
         userRole === UserRoleType.Moderator
       );
       onDeletePost(tx);
@@ -217,7 +213,7 @@ export function PostContent(props: PostContentProps) {
         {showGiveAward && postToAward && (
           <GiveAward
             post={postToAward}
-            collectionId={forumData.info.collectionId}
+            collectionId={forumData.collectionId}
             onCancel={() => setShowGiveAward(false)}
             onSuccess={(notificationContent) => {
               setShowGiveAward(false);