+++ src/components/forums/topic/PostsList.tsx
@@ -8,6 +8,7 @@ import { DispatchForum } from "../../../utils/postbox/postboxWrapper";
 import { UserRoleType } from "../../../utils/permissions";
 import { ForumData } from '../../../utils/hooks';
 import { selectRepliesFromPosts } from '../../../utils/posts';
+import { isSuccess } from '../../../utils/loading';
 
 interface PostListProps {
   forum: DispatchForum;