+++ src/views/ForumView/index.tsx
@@ -29,8 +29,18 @@ import { selectTopics } from "../../utils/posts";
 import { useForum, useRole } from "./../../contexts/DispatchProvider";
 import { newPublicKey } from "./../../utils/postbox/validateNewPublicKey";
 import { getUserRole } from "./../../utils/postbox/userRole";
-import { Loading } from "../../types/loading";
-import { useForumData } from "../../utils/hooks";
+import { Loading } from '../../types/loading';
+import {
+  isSuccess,
+  isInitial,
+  isPending,
+  isNotFound,
+  isDispatchClientError
+} from '../../utils/loading';
+import {
+  useForumData,
+  useModal
+} from '../../utils/hooks';
 
 interface ForumViewProps {
   collectionId: string;
@@ -72,13 +82,9 @@ export const ForumView = (props: ForumViewProps) => {
   const { isNotEmpty, wallet, permission } = forumObject;
   const { publicKey } = wallet;
 
-  const [modalInfo, setModalInfo] = useState<{
-    title: string | ReactNode;
-    type: MessageType;
-    body?: string | ReactNode;
-    collapsible?: CollapsibleProps;
-  } | null>(null);
   const [croppedCollectionID, setCroppedCollectionId] = useState<string>("");
+
+  const { modal, showModal } = useModal();
 
   const collectionId = props.collectionId;
   const collectionPublicKey = useMemo(() => {
@@ -95,7 +101,7 @@ export const ForumView = (props: ForumViewProps) => {
     } catch (error) {
       const message = JSON.stringify(error);
       console.log(error);
-      setModalInfo({
+      showModal({
         title: "Something went wrong!",
         type: MessageType.error,
         body: "Invalid Collection ID Public Key",
@@ -120,7 +126,7 @@ export const ForumView = (props: ForumViewProps) => {
     if (isNotEmpty) {
       createForum();
     } else {
-      setModalInfo({
+      showModal({
         title: "Something went wrong",
         type: MessageType.warning,
         body: "Connect to your wallet in order to create a forum",
@@ -133,7 +139,7 @@ export const ForumView = (props: ForumViewProps) => {
 
     try {
       if (!wallet) {
-        setModalInfo({
+        showModal({
           title: "Something went wrong!",
           type: MessageType.error,
           body: `The forum '${title}' for the collection ${croppedCollectionID} could not be created.`,
@@ -167,7 +173,7 @@ export const ForumView = (props: ForumViewProps) => {
         }
 
         setShowNewForumModal(false);
-        setModalInfo({
+        showModal({
           title: `Success!`,
           body: (
             <div className="successBody">
@@ -188,7 +194,7 @@ export const ForumView = (props: ForumViewProps) => {
         setShowNewForumModal(true);
       } else {
         setShowNewForumModal(false);
-        setModalInfo({
+        showModal({
           title: "Something went wrong!",
           type: MessageType.error,
           body: `The forum '${title}' for the collection ${croppedCollectionID} could not be created.`,
@@ -202,15 +208,14 @@ export const ForumView = (props: ForumViewProps) => {
 
   useEffect(() => {
     update();
-    // Update every time wallet or cluster is changed
-  }, [forumObject.wallet, forumObject.cluster]);
+    // Update every time the cluster is changed
+  }, [forumObject.cluster]);
 
   useEffect(() => {
-    if (
+    if(
       isNotEmpty &&
-      forumData.state === "success" &&
-      forumObject.wallet.publicKey
-    ) {
+      isSuccess(forumData) &&
+      forumObject.wallet.publicKey) {
       getUserRole(forumObject, collectionPublicKey!, Role);
     }
   }, [forumData, isNotEmpty, publicKey]);
@@ -225,7 +230,7 @@ export const ForumView = (props: ForumViewProps) => {
           if (isNotEmpty) {
             setShowNewForumModal(true);
           } else {
-            setModalInfo({
+            showModal({
               title: "Something went wrong",
               type: MessageType.warning,
               body: "Connect to your wallet in order to create a forum",
@@ -262,21 +267,7 @@ export const ForumView = (props: ForumViewProps) => {
   return (
     <div className="dsp-">
       <div className="forumView">
-        {!_.isNil(modalInfo) && (
-          <PopUpModal
-            id="create-forum-info"
-            visible
-            title={modalInfo.title}
-            messageType={modalInfo.type}
-            body={modalInfo.body}
-            collapsible={modalInfo.collapsible}
-            okButton={
-              <a className="okInfoButton" onClick={() => setModalInfo(null)}>
-                OK
-              </a>
-            }
-          />
-        )}
+        {modal}
         {showNewForumModal && (
           <PopUpModal
             id="create-forum"
@@ -353,44 +344,46 @@ export const ForumView = (props: ForumViewProps) => {
         {!permission.readAndWrite && <ConnectionAlert />}
         <div className="forumViewContainer">
           <div className="forumViewContent">
-            {forumData.state === "success" && (
+
+            {isSuccess(forumData) &&
+              (
               <div
                 className={`forumViewTitle ${
                   !permission.readAndWrite ? "alert" : ""
-                }`}
-              >
-                {forumData.value.info.title}
-                <title>{forumData.value.info.title} Forum</title>
+                }`}>
+                {forumData.description.title}
+                <title>{forumData.description.title} Forum</title>
                 <meta
                   name="description"
-                  content={forumData.value.info.description}
+                  content={forumData.description.desc}
                 />
               </div>
-            )}
+              ) /* TODO(andrew) what to render here if title isn't loaded */}
             <main>
               <div className="forumViewContentBox">
                 <div>
                   {(() => {
-                    if (
-                      forumData.state === "initial" ||
-                      forumData.state === "pending"
-                    ) {
+                    if (isSuccess(forumData)) {
+                      return (
+                        <ForumContent
+                          forumObject={forumObject}
+                          forumData={forumData}
+                          update={update}
+                        />
+                      );
+                    } else if (
+                      isInitial(forumData) ||
+                      isPending(forumData)) {
                       return (
                         <div className="forumLoading">
                           <Spinner />
                         </div>
                       );
-                    } else if (forumData.state === "success") {
-                      return (
-                        <ForumContent
-                          forumObject={forumObject}
-                          forumData={forumData.value}
-                          update={update}
-                        />
-                      );
-                    } else if (forumData.state === "notFound") {
+                    } else if(isNotFound(forumData)) {
                       return emptyView;
-                    } else if (forumData.state === "failed") {
+                    } else {
+                      // TODO(andrew) better, more detailed error
+                      // view here
                       return disconnectedView;
                     }
                   })()}